bookstoremanagementSystem/pythondjango/html/css
The authorized employees of bookstore portals upload all details about each book product components in to the server such as book id, book name, author name, version details, and rate and tax details also. The customers who want to access this resource should be registered and get login credential. In registration process the customer type such as single customer or retailer is mandatory.  The concern fixes the rate and tax for the products based on the customer type.This information is retrieved and used for billing module.  In a billing section, the sold product details, no of items sold, rate of the product, concession amount based on the customer type and total bill amount details are stored to a server and an invoice also generated. The authorized administrator views the customer details, product stock information and also able to view the report of the products. In this report the frequently sold company products are highlighted and which customers buy many products are also highlighted. Based on this report the admin balances the stock, maintain customer relationship such as give offer, rate reduction for a valuable customer.